Remove special cxxflags setting for shlib 'halco_common'
When building haldls as a spack package the system header vs. normal header
detection does something different than in our typical, use-container-view,
builds. As cereal containes an old version of rapidjson (which itself triggers
warnings (since gcc 7)) the build subsequently fails in the container.
This changeset shall workaround this issue (and the occurance of random
compiler options in wscripts is bad anyways).

Change-Id: Ia904327ef246a802f5865c318f9667af0e8fa163